services/pedidos/requirements.txt
Flask==1.0.2
Flask-CLI==0.4.0
Flask-SQLAlchemy==2.3.2

python-dateutil